+++ extension/src/shared/view/fields/general/LayerLayerDescriptionField.tsx
@@ -1,19 +1,11 @@
-import { styled } from "@mui/material";
 import { useAtom } from "jotai";
 import { FC } from "react";
 import Markdown from "react-markdown";
 
 import { ParameterList } from "../../../../prototypes/ui-components";
 import { LayerDescriptionField } from "../../../types/fieldComponents/general";
+import { CommonContentWrapper } from "../../../ui-components/CommonContentWrapper";
 import { WritableAtomForComponent } from "../../../view-layers/component";
-
-const Wrapper = styled("div")(({ theme }) => ({
-  padding: theme.spacing(1),
-  fontSize: theme.typography.body2.fontSize,
-  [`img`]: {
-    maxWidth: "100%",
-  },
-}));
 
 export interface LayerLayerDescriptionFieldProps {
   atoms: WritableAtomForComponent<LayerDescriptionField>[];
@@ -23,9 +15,9 @@ export const LayerLayerDescriptionField: FC<LayerLayerDescriptionFieldProps> = (
   const [component] = useAtom(atoms[0]);
   return component.preset?.description ? (
     <ParameterList>
-      <Wrapper>
+      <CommonContentWrapper>
         <Markdown skipHtml>{component.preset?.description}</Markdown>
-      </Wrapper>
+      </CommonContentWrapper>
     </ParameterList>
   ) : null;
 };
